+++ src/components/Timeline.tsx
@@ -3,17 +3,26 @@ import StyledTimeline from '@styles/components/Timeline.styled';
 
 type TimelineProps = { children: React.ReactNode };
 
-type DayProps = { day: number }
+type DayProps = {
+  day: number;
+  recordCount?: number;
+};
 
 type TimelineComposition = { Day: React.FC<DayProps> };
 
-const Day: React.FC<DayProps & {}> = ({ day }: DayProps): JSX.Element => {
+const Day: React.FC<DayProps & {}> = ({ day, recordCount }: DayProps): JSX.Element => {
   const getDayString = (): string => {
     const date = new Date();
     date.setDate(date.getDate() - day);
     return `${date.getMonth() + 1}/${date.getDate()}/${date.getFullYear()}`;
   };
-  return <div className='day'>{getDayString()}</div>;
+  return (
+    <div className='day'>
+      { getDayString() }
+      <br />
+      { recordCount }
+    </div>
+  );
 };
 
 const Timeline: React.FC<TimelineProps> & TimelineComposition =